refactor: without context
